refactor(security): promote KeyRotationPolicy to spi.security.identity (v0.10 S1, ADR-040)
Moves the format-blind rotation-timing carrier from eu.exeris.kernel.community.security
to eu.exeris.kernel.spi.security.identity so it sits in the IdentityProvider SPI seam
(ADR-040) — future drivers and the Enterprise overlay share one policy type. Introduced
in Community at 0.9.0; @since on the SPI location is 0.10.0.

- New: spi.security.identity.KeyRotationPolicy (record + defaults(), content unchanged).
- Removed: community.security.KeyRotationPolicy.
- Repointed imports: CommunityRotatingKeySet + CommunitySecurityProviderTckTest +
  CommunityJwksKeyRotationJfrTest.

Behaviour unchanged: community security tests 45/0/0 (TCK + JWKS-rotation JFR green);
pmd + checkstyle clean.
